Re-introduced lifetimes into Datalog AST
This is nice, because we don't have this stupid `'static` requirement anymore, which turns out to be prone to bugs due to `Span`'s `source_ref()` method.

Dodged a lot of lifetime issues by realising we were over-mutating state; in maps, we only needed the map itself to be mutable, not its elements.