Working on implementing the prototype environment for new ontology
So far quite nice. Hitting a little bump by not being able to implement `From<Message> for MessageSet` for any reference-like (& or Cow) because of either 1) it being a reference or 2) it would require an foreign-impl-on-foreign-type.

Solution: probably move the `From<_>` into the Message. _Or_ make it not a strict requirement. Actually maybe that's nicer.