It aaaaaalmost works
There's this one problem of temporarily referring to the policy at validation time. I can't get it to work; there seems to be a tension between needing to mention the iterator to `ExtractPolicy` at trait level in order to impose custom bounds on the toplevel; yet an inability to do so because we can't really name the temporary in-body lifetime.

Guess we'll have to keep trying how to work this out.