wah
OK, so, this is pretty much impossible. No matter which way I turn, there won't be any way to express _generically_ that there is some policy referring to something that is beyond a `MessageSet`. Once again, it's the some-message-sets-own-their-data, some-message-sets-refer-their-data reference problem.

I might write a blogpost about this at sometime, though. Because this keeps frickin' haunting me xD