Implemented dictator schemes for Agreements and Times
Also finished the Statements. Bit of an unfortunate one that we can't really Rc<RefCell<>> it due to lifetimes; turns out we might need a different type of set after all. For now, it'll work by simply handing control back to the simulator that processes the results. That won't work in an actually distributed scenario, though...